Fix preview using ID instead of email
FossilOrigin-Name: 40d703fd3f9decfc94118f54bea26653acb1662e4cc49759204c3b85d114692b

+++ src/www/admin/users/mailing/details.php
@@ -13,7 +13,7 @@ if (!$mailing) {
 }
 
 if (qg('preview') !== null) {
-	echo $mailing->getHTMLPreview(qg('preview') ?: null, true);
+	echo $mailing->getHTMLPreview((int)qg('preview') ?: null, true);
 	return;
 }
